feat: add RelatedStories component and implement story utility helpers for text parsing and file exports
