+++ frontend/src/components/stories/RelatedStories.tsx
@@ -1,0 +1,36 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { IStories } from "./stories.view.component";
+
+export interface IRelatedStoriesComponentProps {
+  posts: (IStories & { _id: string })[];
+  currentPostId: string;
+}
+
+export const RelatedStoriesComponent: React.FC<IRelatedStoriesComponentProps> = ({ posts, currentPostId }) => {
+  const navigate = useNavigate();
+  const filteredPosts = posts.filter((post) => post._id !== currentPostId);
+
+  return (
+    <div className="mt-8">
+      <h4 className="text-lg font-bold text-slate-200 mb-4">Related Content</h4>
+      {filteredPosts.length > 0 ? (
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+          {filteredPosts.map((post) => (
+            <div
+              key={post._id}
+              onClick={() => navigate(`/stories/${post._id}`)}
+              className="p-4 bg-slate-700/40 rounded-xl border border-slate-600/30 cursor-pointer hover:bg-slate-700/60 transition-colors"
+            >
+              <p className="text-sm font-semibold text-white truncate">{post.title}</p>
+            </div>
+          ))}
+        </div>
+      ) : (
+        <p className="text-center text-slate-500 py-4 border border-dashed border-slate-700 rounded-xl">No related stories found.</p>
+      )}
+    </div>
+  );
+};
+
+export default RelatedStoriesComponent;
